Add types and organize better
