+++ src/types/PurchaseCategoryCount.ts
@@ -1,0 +1,4 @@
+export interface CategoryCount {
+  category: string;
+  count: number;
+}
